[Durable Agents] Split temporal activities for agent loop
- Move runModelActivity and runToolActivity to separate files
- Remove old dummy activities.ts file
- Update imports in worker and workflows
- This completes the migration to temporal activities pattern
