+++ front/lib/api/assistant/agent.ts
@@ -65,82 +65,10 @@ import type {
   ReasoningContentType,
   TextContentType,
 } from "@app/types/assistant/agent_message_content";
+import { runModelActivity } from "@app/temporal/agent_loop/activities/run_model";
+import { runToolActivity } from "@app/temporal/agent_loop/activities/run_tool";
 
-const CANCELLATION_CHECK_INTERVAL = 500;
 const MAX_ACTIONS_PER_STEP = 16;
-const MAX_AUTO_RETRY = 3;
-
-// Process database operations for agent events before publishing to Redis.
-async function processEventForDatabase(
-  event: AgentMessageEvents,
-  agentMessageRow: AgentMessage,
-  step: number
-): Promise<void> {
-  switch (event.type) {
-    case "agent_error":
-      // Store error in database.
-      await agentMessageRow.update({
-        status: "failed",
-        errorCode: event.error.code,
-        errorMessage: event.error.message,
-        errorMetadata: event.error.metadata,
-      });
-      await AgentStepContentResource.makeNew({
-        workspaceId: agentMessageRow.workspaceId,
-        agentMessageId: agentMessageRow.id,
-        step,
-        index: 0, // Errors are the only content for this step
-        type: "error",
-        value: {
-          type: "error",
-          value: {
-            code: event.error.code,
-            message: event.error.message,
-            metadata: {
-              ...event.error.metadata,
-              category: event.error.metadata?.category || "",
-            },
-          },
-        },
-        version: 0,
-      });
-      break;
-
-    case "agent_generation_cancelled":
-      // Store cancellation in database.
-      await agentMessageRow.update({
-        status: "cancelled",
-      });
-      break;
-
-    case "agent_message_success":
-      // Store success and run IDs in database.
-      await agentMessageRow.update({
-        runIds: event.runIds,
-        status: "succeeded",
-      });
-
-      break;
-
-    default:
-      // Ensure we handle all event types.
-      break;
-  }
-}
-
-async function updateResourceAndPublishEvent(
-  event: AgentMessageEvents,
-  conversation: ConversationType,
-  agentMessageRow: AgentMessage,
-  step: number
-): Promise<void> {
-  // Process database operations BEFORE publishing to Redis.
-  await processEventForDatabase(event, agentMessageRow, step);
-
-  await publishConversationRelatedEvent(event, {
-    conversationId: conversation.sId,
-  });
-}
 
 // This interface is used to execute an agent. It is not in charge of creating the AgentMessage,
 // but it now handles updating it based on the execution results.
@@ -160,7 +88,7 @@ export async function runAgentWithStreaming(
 
   await wakeLock(async () => {
     for (let i = 0; i < MAX_STEPS_USE_PER_RUN_LIMIT + 1; i++) {
-      const result = await runMultiActionsAgent({
+      const result = await runModelActivity({
         authType,
         runAgentArgs,
         runIds,
@@ -190,7 +118,7 @@ export async function runAgentWithStreaming(
             ? functionCallStepContentIds[functionCallId]
             : undefined;
 
-          return runAction(authType, {
+          return runToolActivity(authType, {
             runAgentArgs,
             inputs,
             functionCallId,
@@ -219,948 +147,4 @@ export async function runAgentWithStreaming(
   await launchUpdateUsageWorkflow({
     workspaceId: authType.workspaceId,
   });
-}
-
-// This method is used by the multi-actions execution loop to pick the next
-// action to execute and generate its inputs.
-//
-// TODO(DURABLE-AGENTS 2025-07-20): The method mutates agentMessage, this must
-// be refactored in a follow up PR.
-async function runMultiActionsAgent({
-  authType,
-  runAgentArgs,
-  runIds,
-  step,
-  functionCallStepContentIds,
-  autoRetryCount = 0,
-}: {
-  authType: AuthenticatorType;
-  runAgentArgs: RunAgentArgs;
-  runIds: string[];
-  step: number;
-  functionCallStepContentIds: Record<string, ModelId>;
-  autoRetryCount?: number;
-}): Promise<{
-  actions: AgentActionsEvent["actions"];
-  runId: string;
-  functionCallStepContentIds: Record<string, ModelId>;
-} | null> {
-  const {
-    agentConfiguration,
-    conversation,
-    userMessage,
-    agentMessage,
-    agentMessageRow,
-  } = getRunAgentData(runAgentArgs);
-
-  const now = Date.now();
-
-  const localLogger = logger.child({
-    workspaceId: conversation.owner.sId,
-    conversationId: conversation.sId,
-    multiActionLoopIteration: step,
-  });
-
-  localLogger.info("Starting multi-action loop iteration");
-
-  const isLegacyAgent = isLegacyAgentConfiguration(agentConfiguration);
-  if (isLegacyAgent && step !== 0) {
-    // legacy agents stop after one step
-    return null;
-  }
-
-  const isLastGenerationIteration = step === agentConfiguration.maxStepsPerRun;
-
-  const agentActions =
-    // If we already executed the maximum number of actions, we don't run anymore.
-    // This will force the agent to run the generation.
-    isLastGenerationIteration
-      ? []
-      : // Otherwise, we let the agent decide which action to run (if any).
-        agentConfiguration.actions;
-
-  const auth = await Authenticator.fromJSON(authType);
-
-  const model = getSupportedModelConfig(agentConfiguration.model);
-
-  // Helper function to publish agent error events
-  async function publishAgentError(error: {
-    code: string;
-    message: string;
-    metadata: Record<string, string | number | boolean> | null;
-  }): Promise<void> {
-    // Check if this is a multi_actions_error that hit max retries
-    let logMessage = `Agent error: ${error.message}`;
-    if (
-      error.code === "multi_actions_error" &&
-      error.metadata?.retriesAttempted === MAX_AUTO_RETRY
-    ) {
-      logMessage = `Agent error: ${error.message} (max retries reached)`;
-    } else if (
-      error.code === "multi_actions_error" &&
-      error.metadata?.category &&
-      error.metadata.category !== "retryable_model_error"
-    ) {
-      logMessage = `Agent error: ${error.message} (not retryable)`;
-    }
-
-    localLogger.error(
-      {
-        error,
-      },
-      logMessage
-    );
-
-    await updateResourceAndPublishEvent(
-      {
-        type: "agent_error",
-        created: Date.now(),
-        configurationId: agentConfiguration.sId,
-        messageId: agentMessage.sId,
-        error,
-      },
-      conversation,
-      agentMessageRow,
-      step
-    );
-  }
-
-  // Helper function to flush all pending tokens from the content parser
-  async function flushParserTokens(): Promise<void> {
-    for await (const tokenEvent of contentParser.flushTokens()) {
-      await updateResourceAndPublishEvent(
-        tokenEvent,
-        conversation,
-        agentMessageRow,
-        step
-      );
-    }
-  }
-
-  if (!model) {
-    await publishAgentError({
-      code: "model_does_not_support_multi_actions",
-      message:
-        `The model you selected (${agentConfiguration.model.modelId}) ` +
-        `does not support multi-actions.`,
-      metadata: null,
-    });
-    return null;
-  }
-  const availableActions: ActionConfigurationType[] = [];
-
-  for (const agentAction of agentActions) {
-    if (isActionConfigurationType(agentAction)) {
-      availableActions.push(agentAction);
-    }
-  }
-
-  const attachments = listAttachments(conversation);
-  const jitServers = await getJITServers(auth, {
-    conversation,
-    attachments,
-  });
-
-  // Get client-side MCP server configurations from user message context.
-  const clientSideMCPActionConfigurations =
-    await createClientSideMCPServerConfigurations(
-      auth,
-      userMessage.context.clientSideMCPServerIds
-    );
-
-  const {
-    serverToolsAndInstructions: mcpActions,
-    error: mcpToolsListingError,
-  } = await tryListMCPTools(
-    auth,
-    {
-      agentConfiguration,
-      conversation,
-      agentMessage,
-      clientSideActionConfigurations: clientSideMCPActionConfigurations,
-    },
-    jitServers
-  );
-
-  if (mcpToolsListingError) {
-    localLogger.error(
-      {
-        error: mcpToolsListingError,
-      },
-      "Error listing MCP tools."
-    );
-  }
-
-  if (!isLastGenerationIteration) {
-    availableActions.push(...mcpActions.flatMap((s) => s.tools));
-  }
-
-  let fallbackPrompt = "You are a conversational agent";
-  if (
-    agentConfiguration.actions.length ||
-    agentConfiguration.visualizationEnabled ||
-    availableActions.length > 0
-  ) {
-    fallbackPrompt += " with access to tool use.";
-  } else {
-    fallbackPrompt += ".";
-  }
-
-  const agentsList = agentConfiguration.instructions?.includes(
-    "{ASSISTANTS_LIST}"
-  )
-    ? await getAgentConfigurations({
-        auth,
-        agentsGetView: auth.user() ? "list" : "all",
-        variant: "light",
-      })
-    : null;
-
-  const prompt = await constructPromptMultiActions(auth, {
-    userMessage,
-    agentConfiguration,
-    fallbackPrompt,
-    model,
-    hasAvailableActions: availableActions.length > 0,
-    errorContext: mcpToolsListingError,
-    agentsList,
-    conversationId: conversation.sId,
-    serverToolsAndInstructions: mcpActions,
-  });
-
-  const specifications: AgentActionSpecification[] = [];
-  for (const a of availableActions) {
-    const specRes =
-      await getRunnerForActionConfiguration(a).buildSpecification(auth);
-
-    if (specRes.isErr()) {
-      await publishAgentError({
-        code: "build_spec_error",
-        message: `Failed to build the specification for action ${a.sId},`,
-        metadata: null,
-      });
-
-      return null;
-    }
-
-    // Truncate the description to 1024 characters
-    specRes.value.description = specRes.value.description.slice(0, 1024);
-
-    specifications.push(specRes.value);
-  }
-
-  // Count the number of tokens used by the functions presented to the model.
-  // This is a rough estimate of the number of tokens.
-  const tools = JSON.stringify(
-    specifications.map((s) => ({
-      name: s.name,
-      description: s.description,
-      inputSchema: s.inputSchema,
-    }))
-  );
-
-  // Turn the conversation into a digest that can be presented to the model.
-  const modelConversationRes = await renderConversationForModel(auth, {
-    conversation,
-    model,
-    prompt,
-    tools,
-    allowedTokenCount: model.contextSize - model.generationTokensCount,
-  });
-
-  if (modelConversationRes.isErr()) {
-    await publishAgentError({
-      code: "conversation_render_error",
-      message: `Error rendering conversation for model: ${modelConversationRes.error.message}`,
-      metadata: null,
-    });
-
-    return null;
-  }
-
-  // Check that specifications[].name are unique. This can happen if the user overrides two actions
-  // names with the same name (advanced settings). We return an actionable error if that's the case
-  // as we want to keep that as an invariant when interacting with models.
-  const seen = new Set<string>();
-  for (const spec of specifications) {
-    if (seen.has(spec.name)) {
-      await publishAgentError({
-        code: "duplicate_specification_name",
-        message:
-          `Duplicate action name in agent configuration: ${spec.name}. ` +
-          "Your agents actions must have unique names.",
-        metadata: null,
-      });
-
-      return null;
-    }
-    seen.add(spec.name);
-  }
-
-  const runConfig = cloneBaseConfig(
-    getDustProdAction("assistant-v2-multi-actions-agent").config
-  );
-  if (isLegacyAgent) {
-    runConfig.MODEL.function_call =
-      specifications.length === 1 ? specifications[0].name : null;
-  } else {
-    runConfig.MODEL.function_call = specifications.length === 0 ? null : "auto";
-  }
-  runConfig.MODEL.provider_id = model.providerId;
-  runConfig.MODEL.model_id = model.modelId;
-  runConfig.MODEL.temperature = agentConfiguration.model.temperature;
-
-  const reasoningEffort =
-    agentConfiguration.model.reasoningEffort ?? model.defaultReasoningEffort;
-  if (reasoningEffort !== "none" && reasoningEffort !== "light") {
-    runConfig.MODEL.reasoning_effort = reasoningEffort;
-  }
-
-  if (agentConfiguration.model.responseFormat) {
-    runConfig.MODEL.response_format = JSON.parse(
-      agentConfiguration.model.responseFormat
-    );
-  }
-  const anthropicBetaFlags = config.getMultiActionsAgentAnthropicBetaFlags();
-  if (anthropicBetaFlags) {
-    runConfig.MODEL.anthropic_beta_flags = anthropicBetaFlags;
-  }
-
-  const res = await runActionStreamed(
-    auth,
-    "assistant-v2-multi-actions-agent",
-    runConfig,
-    [
-      {
-        conversation: modelConversationRes.value.modelConversation,
-        specifications,
-        prompt,
-      },
-    ],
-    {
-      conversationId: conversation.sId,
-      workspaceId: conversation.owner.sId,
-      userMessageId: userMessage.sId,
-    }
-  );
-
-  // Errors occurring during the multi-actions-agent dust app may be retryable.
-  // Their implicit code should be "multi_actions_error".
-  async function handlePossiblyRetryableError(message: string) {
-    const { category, publicMessage, errorTitle } = categorizeAgentErrorMessage(
-      {
-        code: "multi_actions_error",
-        message,
-      }
-    );
-
-    const isRetryableModelError = [
-      "retryable_model_error",
-      "stream_error",
-    ].includes(category);
-
-    if (isRetryableModelError && autoRetryCount < MAX_AUTO_RETRY) {
-      localLogger.warn(
-        {
-          error: message,
-          retryCount: autoRetryCount + 1,
-          maxRetries: MAX_AUTO_RETRY,
-        },
-        "Auto-retrying multi-actions agent due to retryable model error."
-      );
-
-      // Recursively retry with incremented count
-      return runMultiActionsAgent({
-        authType,
-        runAgentArgs,
-        runIds,
-        step,
-        functionCallStepContentIds,
-        autoRetryCount: autoRetryCount + 1,
-      });
-    }
-
-    await publishAgentError({
-      code: "multi_actions_error",
-      message: publicMessage,
-      metadata: {
-        category,
-        errorTitle,
-        retriesAttempted: autoRetryCount,
-      },
-    });
-
-    return null;
-  }
-
-  if (res.isErr()) {
-    return handlePossiblyRetryableError(res.error.message);
-  }
-
-  const { eventStream, dustRunId } = res.value;
-  let output: {
-    actions: Array<{
-      functionCallId: string | null;
-      name: string | null;
-      arguments: Record<string, string | boolean | number> | null;
-    }>;
-    generation: string | null;
-    contents: Array<
-      TextContentType | FunctionCallContentType | ReasoningContentType
-    >;
-  } | null = null;
-
-  let shouldYieldCancel = false;
-  let lastCheckCancellation = Date.now();
-  const redis = await getRedisClient({ origin: "assistant_generation" });
-  let isGeneration = true;
-
-  const contentParser = new AgentMessageContentParser(
-    agentConfiguration,
-    agentMessage.sId,
-    getDelimitersConfiguration({ agentConfiguration })
-  );
-
-  async function checkCancellation() {
-    try {
-      const cancelled = await redis.get(
-        `assistant:generation:cancelled:${agentMessage.sId}`
-      );
-      if (cancelled === "1") {
-        shouldYieldCancel = true;
-        await redis.set(
-          `assistant:generation:cancelled:${agentMessage.sId}`,
-          0,
-          {
-            EX: 3600, // 1 hour
-          }
-        );
-      }
-    } catch (error) {
-      localLogger.error({ error }, "Error checking cancellation");
-      return false;
-    }
-  }
-
-  let nativeChainOfThought = "";
-
-  // Create a new object to avoid mutation
-  const updatedFunctionCallStepContentIds = { ...functionCallStepContentIds };
-
-  for await (const event of eventStream) {
-    if (event.type === "function_call") {
-      isGeneration = false;
-    }
-
-    if (event.type === "error") {
-      await flushParserTokens();
-      return handlePossiblyRetryableError(event.content.message);
-    }
-
-    const currentTimestamp = Date.now();
-    if (
-      currentTimestamp - lastCheckCancellation >=
-      CANCELLATION_CHECK_INTERVAL
-    ) {
-      void checkCancellation(); // Trigger the async function without awaiting
-      lastCheckCancellation = currentTimestamp;
-    }
-
-    if (shouldYieldCancel) {
-      await flushParserTokens();
-      await updateResourceAndPublishEvent(
-        {
-          type: "agent_generation_cancelled",
-          created: Date.now(),
-          configurationId: agentConfiguration.sId,
-          messageId: agentMessage.sId,
-        },
-        conversation,
-        agentMessageRow,
-        step
-      );
-      return null;
-    }
-
-    if (event.type === "tokens" && isGeneration) {
-      for await (const tokenEvent of contentParser.emitTokens(
-        event.content.tokens.text
-      )) {
-        await updateResourceAndPublishEvent(
-          tokenEvent,
-          conversation,
-          agentMessageRow,
-          step
-        );
-      }
-    }
-
-    if (event.type === "reasoning_tokens") {
-      await updateResourceAndPublishEvent(
-        {
-          type: "generation_tokens",
-          classification: "chain_of_thought",
-          created: Date.now(),
-          configurationId: agentConfiguration.sId,
-          messageId: agentMessage.sId,
-          text: event.content.tokens.text,
-        },
-        conversation,
-        agentMessageRow,
-        step
-      );
-      nativeChainOfThought += event.content.tokens.text;
-    }
-
-    if (event.type === "reasoning_item") {
-      await updateResourceAndPublishEvent(
-        {
-          type: "generation_tokens",
-          classification: "chain_of_thought",
-          created: Date.now(),
-          configurationId: agentConfiguration.sId,
-          messageId: agentMessage.sId,
-          text: "\n\n",
-        },
-        conversation,
-        agentMessageRow,
-        step
-      );
-      nativeChainOfThought += "\n\n";
-    }
-
-    if (event.type === "block_execution") {
-      const e = event.content.execution[0][0];
-      if (e.error) {
-        await flushParserTokens();
-        return handlePossiblyRetryableError(e.error);
-      }
-
-      if (event.content.block_name === "MODEL" && e.value) {
-        // Flush early as we know the generation is terminated here.
-        await flushParserTokens();
-
-        const block = e.value;
-        if (!isDustAppChatBlockType(block)) {
-          return handlePossiblyRetryableError(
-            "Received unparsable MODEL block."
-          );
-        }
-
-        // Extract token usage from block execution metadata
-        const meta = e.meta as {
-          token_usage?: {
-            prompt_tokens: number;
-            completion_tokens: number;
-            reasoning_tokens?: number;
-          };
-        } | null;
-        const reasoningTokens = meta?.token_usage?.reasoning_tokens || 0;
-
-        const contents = (block.message.contents ?? []).map((content) => {
-          if (content.type === "reasoning") {
-            return {
-              ...content,
-              value: {
-                ...content.value,
-                tokens: 0, // Will be updated for the last reasoning item
-                provider: model.providerId,
-              },
-            } satisfies ReasoningContentType;
-          }
-          return content;
-        });
-
-        // We unfortunately don't currently have a proper breakdown of reasoning tokens per item,
-        // so we set the reasoning token count on the last reasoning item.
-        for (let i = contents.length - 1; i >= 0; i--) {
-          const content = contents[i];
-          if (content.type === "reasoning") {
-            content.value.tokens = reasoningTokens;
-            contents[i] = content;
-            break;
-          }
-        }
-
-        output = {
-          actions: [],
-          generation: null,
-          contents,
-        };
-
-        if (block.message.function_calls?.length) {
-          for (const fc of block.message.function_calls) {
-            try {
-              const args = JSON.parse(fc.arguments);
-              output.actions.push({
-                name: fc.name,
-                functionCallId: fc.id,
-                arguments: args,
-              });
-            } catch (error) {
-              await publishAgentError({
-                code: "function_call_error",
-                message: `Error parsing function call arguments: ${error}`,
-                metadata: null,
-              });
-              return null;
-            }
-          }
-        } else {
-          output.generation = block.message.content ?? null;
-        }
-      }
-    }
-  }
-
-  await flushParserTokens();
-
-  if (!output) {
-    return handlePossiblyRetryableError("Agent execution didn't complete.");
-  }
-
-  // Create AgentStepContent for each content item (reasoning, text, function calls)
-  // This replaces the original agent_step_content event emission
-  for (const [index, content] of output.contents.entries()) {
-    const stepContent = await AgentStepContentResource.makeNew({
-      workspaceId: conversation.owner.id,
-      agentMessageId: agentMessage.agentMessageId,
-      step,
-      index,
-      version: autoRetryCount,
-      type: content.type,
-      value: content,
-    });
-
-    // If this is a function call content, track the step content ID
-    if (content.type === "function_call") {
-      updatedFunctionCallStepContentIds[content.value.id] = stepContent.id;
-    }
-  }
-
-  // Track retries that lead to completing successfully (with either function calls or generation).
-  if (autoRetryCount > 0) {
-    statsDClient.increment("successful_auto_retry.count", 1, [
-      `retryCount:${autoRetryCount}`,
-    ]);
-  }
-
-  // Store the contents for returning to the caller
-  // These will be added to agentMessage.contents in the calling function
-
-  if (!output.actions.length) {
-    // Successful generation.
-    const processedContent = contentParser.getContent() ?? "";
-    if (!processedContent.length) {
-      localLogger.warn(
-        {
-          modelId: model.modelId,
-        },
-        "No content generated by the agent."
-      );
-    }
-
-    // TODO(DURABLE-AGENTS 2025-07-20): Avoid mutating agentMessage here
-    const chainOfThought =
-      (nativeChainOfThought || contentParser.getChainOfThought()) ?? "";
-
-    if (chainOfThought.length) {
-      if (!agentMessage.chainOfThought) {
-        agentMessage.chainOfThought = "";
-      }
-      agentMessage.chainOfThought += chainOfThought;
-    }
-    agentMessage.content = (agentMessage.content ?? "") + processedContent;
-    agentMessage.status = "succeeded";
-
-    await updateResourceAndPublishEvent(
-      {
-        type: "agent_message_success",
-        created: Date.now(),
-        configurationId: agentConfiguration.sId,
-        messageId: agentMessage.sId,
-        message: agentMessage,
-        runIds: [...runIds, await dustRunId],
-      },
-      conversation,
-      agentMessageRow,
-      step
-    );
-
-    return null;
-  }
-
-  // We have actions.
-  localLogger.info(
-    {
-      elapsed: Date.now() - now,
-    },
-    "[ASSISTANT_TRACE] Action inputs generation"
-  );
-
-  if (isLastGenerationIteration) {
-    await publishAgentError({
-      code: "tool_use_limit_reached",
-      message:
-        "The agent attempted to use too many tools. This model error can be safely retried.",
-      metadata: null,
-    });
-    return null;
-  }
-
-  const actions: AgentActionsEvent["actions"] = [];
-
-  for (const a of output.actions) {
-    // Sometimes models will return a name with a triple underscore instead of a double underscore, we dynamically handle it.
-    const actionNamesFromLLM: string[] = removeNulls([
-      a.name,
-      a.name?.replace("___", TOOL_NAME_SEPARATOR) ?? null,
-    ]);
-
-    let action = availableActions.find((ac) =>
-      actionNamesFromLLM.includes(ac.name)
-    );
-    let args = a.arguments;
-
-    if (!action) {
-      if (!a.name) {
-        await publishAgentError({
-          code: "action_not_found",
-          message:
-            `The agent attempted to run an invalid action (no name). ` +
-            `This model error can be safely retried.`,
-          metadata: null,
-        });
-
-        return null;
-      }
-      const mcpServerView =
-        await MCPServerViewResource.getMCPServerViewForAutoInternalTool(
-          auth,
-          "missing_action_catcher"
-        );
-
-      // Could happen if the internal server has not already been added
-      if (!mcpServerView) {
-        await publishAgentError({
-          code: "action_not_found",
-          message:
-            `The agent attempted to run an invalid action (${a.name}). ` +
-            `This model error can be safely retried (no server).`,
-          metadata: null,
-        });
-        return null;
-      }
-
-      localLogger.warn(
-        {
-          actionName: a.name,
-          availableActions: availableActions.map((a) => a.name),
-        },
-        "Model attempted to run an action that is not part of the agent configuration but we'll try to catch it."
-      );
-
-      // Catch-all action.
-      action = {
-        id: -1,
-        sId: generateRandomModelSId(),
-        type: "mcp_configuration" as const,
-        name: a.name,
-        originalName: a.name,
-        description: null,
-        dataSources: null,
-        tables: null,
-        childAgentId: null,
-        reasoningModel: null,
-        timeFrame: null,
-        jsonSchema: null,
-        additionalConfiguration: {},
-        mcpServerViewId: mcpServerView.sId,
-        dustAppConfiguration: null,
-        internalMCPServerId: mcpServerView.internalMCPServerId,
-        inputSchema: {},
-        availability: "auto_hidden_builder",
-        permission: "never_ask",
-        toolServerId: mcpServerView.sId,
-        mcpServerName: "missing_action_catcher" as InternalMCPServerNameType,
-      };
-      args = {};
-    }
-
-    actions.push({
-      action,
-      inputs: args ?? {},
-      functionCallId: a.functionCallId ?? null,
-    });
-  }
-
-  await flushParserTokens();
-
-  const chainOfThought =
-    (nativeChainOfThought || contentParser.getChainOfThought()) ?? "";
-
-  agentMessage.content =
-    (agentMessage.content ?? "") + (contentParser.getContent() ?? "");
-
-  if (chainOfThought.length) {
-    if (!agentMessage.chainOfThought) {
-      agentMessage.chainOfThought = "";
-    }
-    agentMessage.chainOfThought += chainOfThought;
-  }
-
-  const newContents = output.contents.map((content) => ({
-    step,
-    content,
-  }));
-  agentMessage.contents.push(...newContents);
-  return {
-    actions,
-    runId: await dustRunId,
-    functionCallStepContentIds: updatedFunctionCallStepContentIds,
-  };
-}
-
-async function runAction(
-  authType: AuthenticatorType,
-  {
-    runAgentArgs,
-    inputs,
-    functionCallId,
-    step,
-    stepActionIndex,
-    stepActions,
-    citationsRefsOffset,
-    stepContentId,
-  }: {
-    runAgentArgs: RunAgentArgs;
-    inputs: Record<string, string | boolean | number>;
-    functionCallId: string | null;
-    step: number;
-    stepActionIndex: number;
-    stepActions: ActionConfigurationType[];
-    citationsRefsOffset: number;
-    stepContentId?: ModelId;
-  }
-): Promise<{ citationsIncrement: number }> {
-  const auth = await Authenticator.fromJSON(authType);
-
-  const actionConfiguration = stepActions[stepActionIndex];
-  const { agentConfiguration, conversation, agentMessage, agentMessageRow } =
-    getRunAgentData(runAgentArgs);
-
-  if (!stepContentId) {
-    logger.error(
-      {
-        functionCallId,
-        actionConfigurationId: actionConfiguration.sId,
-        agentConfigurationId: agentConfiguration.sId,
-      },
-      "No step content for function call"
-    );
-  }
-
-  if (isMCPToolConfiguration(actionConfiguration)) {
-    const eventStream = getRunnerForActionConfiguration(
-      actionConfiguration
-    ).run(auth, {
-      agentConfiguration: agentConfiguration,
-      conversation,
-      agentMessage,
-      rawInputs: inputs,
-      functionCallId,
-      step,
-      stepActionIndex,
-      stepActions,
-      citationsRefsOffset,
-      stepContentId,
-    });
-
-    for await (const event of eventStream) {
-      switch (event.type) {
-        case "tool_error":
-          await updateResourceAndPublishEvent(
-            {
-              type: "agent_error",
-              created: event.created,
-              configurationId: agentConfiguration.sId,
-              messageId: agentMessage.sId,
-              error: {
-                code: event.error.code,
-                message: event.error.message,
-                metadata: event.error.metadata,
-              },
-            },
-            conversation,
-            agentMessageRow,
-            step
-          );
-          return { citationsIncrement: 0 };
-
-        case "tool_success":
-          await updateResourceAndPublishEvent(
-            {
-              type: "agent_action_success",
-              created: event.created,
-              configurationId: agentConfiguration.sId,
-              messageId: agentMessage.sId,
-              action: event.action,
-            },
-            conversation,
-            agentMessageRow,
-            step
-          );
-
-          // We stitch the action into the agent message. The conversation is expected to include
-          // the agentMessage object, updating this object will update the conversation as well.
-          agentMessage.actions.push(event.action);
-          break;
-
-        case "tool_params":
-        case "tool_approve_execution":
-        case "tool_notification":
-          await updateResourceAndPublishEvent(
-            event,
-            conversation,
-            agentMessageRow,
-            step
-          );
-          break;
-
-        default:
-          assertNever(event);
-      }
-    }
-
-    return {
-      citationsIncrement: getCitationsCount({
-        agentConfiguration: agentConfiguration,
-        stepActions: stepActions,
-        stepActionIndex: stepActionIndex,
-      }),
-    };
-  } else {
-    assertNever(actionConfiguration);
-  }
-}
-
-export async function filterSuggestedNames(
-  owner: WorkspaceType,
-  suggestions: string[] | undefined | null
-) {
-  if (!suggestions || suggestions.length === 0) {
-    return [];
-  }
-  // Filter out suggested names that are already in use in the workspace.
-  const existingNames = (
-    await AgentConfiguration.findAll({
-      where: {
-        workspaceId: owner.id,
-        status: "active",
-      },
-      attributes: ["name"],
-    })
-  ).map((ac) => ac.name.toLowerCase());
-
-  return suggestions?.filter((s) => !existingNames.includes(s.toLowerCase()));
 }
